backend_example/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-auth==2.39.0
idna==3.10
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
requests==2.32.3
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.4.0
stripe==9.1.0
